# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SMART CLINIC - Healthcare SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}/patients/{patientId}/notes:
    get:
      operationId: listConsultationNotes
      tags: [patients]
      summary: List consultation notes for a patient
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of consultation notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsultationNote"
    post:
      operationId: createConsultationNote
      tags: [patients]
      summary: Create a consultation note (doctors only)
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationNoteInput"
      responses:
        "201":
          description: Created consultation note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationNote"
  /clinics/{clinicId}/patients/{patientId}/notes/{noteId}:
    get:
      operationId: getConsultationNote
      tags: [patients]
      summary: Get a specific consultation note
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
        - in: path
          name: noteId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Consultation note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationNote"
    patch:
      operationId: updateConsultationNote
      tags: [patients]
      summary: Update a consultation note
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
        - in: path
          name: noteId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationNoteInput"
      responses:
        "200":
          description: Updated consultation note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationNote"
...
components:
  schemas:
...
    ConsultationNoteInput:
      type: object
      properties:
        chiefComplaint:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        vitalSigns:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        prescription:
          type: ["string", "null"]
        treatmentPlan:
          type: ["string", "null"]
        followUpInstructions:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, pending_lab, pending_pharmacy, completed]
        queueEntryId:
          type: ["string", "null"]
        appointmentId:
          type: ["string", "null"]
    ConsultationNote:
      type: object
      required: [id, clinicId, patientId, doctorId, status, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
        doctorName:
          type: ["string", "null"]
        consultationCode:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, pending_lab, pending_pharmacy, completed]
        queueEntryId:
          type: ["string", "null"]
        appointmentId:
          type: ["string", "null"]
        chiefComplaint:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        vitalSigns:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        prescription:
          type: ["string", "null"]
        treatmentPlan:
          type: ["string", "null"]
        followUpInstructions:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ===== APPOINTMENT SCHEMAS =====
...